1.Use different json file for different action to fix a bug: Only the latest action will be kept if multiple actions share the same json file. for example, A. Publish a wms layer. In the mercurial repository, the json file contains the data to publish a wms layer
B. Empty the gwc for this layer. Now, in the mercurial repository, the json file only have the meta data to
empty gwc.

C. Resynchronize a slave server from the beginning.
we will find that wms layer can't be created and also the empty gwc job
will be failed because of non existed wms layer.

2. Implement a "empty_gwc" feature for publish